+++ src/main/java/com/sparta/myselectshop/exception/GlobalExceptionHandler.java
@@ -1,0 +1,22 @@
+package com.sparta.myselectshop.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+// 예외처리 중앙집중화
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler({IllegalArgumentException.class})
+    public ResponseEntity<RestApiException> handleException(IllegalArgumentException ex) {
+        RestApiException restApiException = new RestApiException(ex.getMessage(), HttpStatus.BAD_REQUEST.value());
+        return new ResponseEntity<>(
+                // HTTP body
+                restApiException,
+                // HTTP status code
+                HttpStatus.BAD_REQUEST
+        );
+    }
+}
